feat(backend): add runner + polish agents + implement typed paths

+++ backend/smtp-agent/src/commonMain/kotlin/util.kt
@@ -2,4 +2,4 @@ package dev.sitar.kmail.smtp.agent
 
 import dev.sitar.kmail.smtp.InternetMessage
 
-val InternetMessage.queueId get() = hashCode().toString(16)
+val InternetMessage.queueId get() = hashCode().toUInt().toString(16)
